requirements.txt
grpcio==1.38.0
grpcio-tools==1.38.0
protobuf==3.17.2
psutil==5.8.0
six==1.16.0
